+++ Cargo.toml
@@ -41,6 +41,7 @@ members = [
   "lib/api/macro-wasmer-engine-test",
   "lib/backend-api",
   "lib/c-api",
+  "lib/c-api-imports",
   "lib/c-api/examples/wasmer-capi-examples-runner",
   "lib/c-api/tests/wasmer-c-api-test-runner",
   "lib/cache",
