+++ lib/c-api/Cargo.toml
@@ -26,6 +26,7 @@ crate-type = ["staticlib", "cdylib", "rlib"]
 # We rename `wasmer` to `wasmer-api` to avoid the conflict with this
 # library name (see `[lib]`).
 wasmer-api = { version = "=7.2.0-rc.1", path = "../api", default-features = false, package = "wasmer" }
+wasmer-c-api-imports = { version = "=7.2.0-rc.1", path = "../c-api-imports", default-features = false }
 wasmer-compiler = { version = "=7.2.0-rc.1", path = "../compiler", optional = true }
 wasmer-compiler-cranelift = { version = "=7.2.0-rc.1", path = "../compiler-cranelift", optional = true }
 wasmer-compiler-llvm = { version = "=7.2.0-rc.1", path = "../compiler-llvm", optional = true }
@@ -65,34 +66,42 @@ wat.workspace = true
 
 [features]
 default = ["wat", "sys-default", "cranelift", "compiler", "wasi", "middlewares"]
-wat = ["wasmer-api/wat"]
+wat = ["wasmer-api/wat", "wasmer-c-api-imports/wat"]
 wasi = ["wasmer-wasix"]
 middlewares = ["compiler", "wasmer-middlewares", "wasmer-api/wasmparser"]
 engine = ["compiler"]
 
 # Enable the `sys` backend.
-sys = ["wasmer-api/sys", "wasmer-compiler"]
+sys = ["wasmer-api/sys", "wasmer-compiler", "wasmer-c-api-imports/sys"]
 
 # Enable the `singlepass` compiler.
 singlepass = [
   "dep:wasmer-compiler-singlepass",
   "wasmer-api/singlepass",
+  "wasmer-c-api-imports/singlepass",
   "compiler",
 ]
 # Enable the `cranelift` compiler. 
 cranelift = [
   "dep:wasmer-compiler-cranelift",
   "wasmer-api/cranelift",
+  "wasmer-c-api-imports/cranelift",
   "compiler",
 ]
 # Enable the `llvm` compiler.
-llvm = ["dep:wasmer-compiler-llvm", "wasmer-api/llvm", "compiler"]
+llvm = [
+  "dep:wasmer-compiler-llvm",
+  "wasmer-api/llvm",
+  "wasmer-c-api-imports/llvm",
+  "compiler",
+]
 
 # Enable the use of compiler. Implies enabling the `sys` feature.
 compiler = [
   "sys",
   "wasmer-compiler",
   "wasmer-api/compiler",
+  "wasmer-c-api-imports/compiler",
   "wasmer-compiler/translator",
   "wasmer-compiler/compiler",
 ]
@@ -103,25 +112,38 @@ compiler-headless = [
   "wasmer-artifact-load",
   "static-artifact-load",
   "wasmer-api/compiler",
+  "wasmer-c-api-imports/compiler",
   "wasmer-compiler/translator",
   "wasmer-compiler/compiler",
 ]
 
 # Enable the `v8` backend.
-v8 = ["wasmer-api/v8"]
+v8 = ["wasmer-api/v8", "wasmer-c-api-imports/v8"]
 
 # Enable the `sys` backend and use it as default. Notice that this does not
 # enable a default compiler, and users should therefore add one (or more) of
 # the `singlepass`, `cranelift` and `llvm` features.
-sys-default = ["sys", "wasmer-api/sys-default"]
+sys-default = ["sys", "wasmer-api/sys-default", "wasmer-c-api-imports/sys-default"]
 
 # Enable the `v8` backend and use it as default.
-v8-default = ["v8", "wasmer-api/v8-default"]
+v8-default = ["v8", "wasmer-api/v8-default", "wasmer-c-api-imports/v8-default"]
 
-wasmer-artifact-load = ["wasmer-compiler/wasmer-artifact-load"]
-wasmer-artifact-create = ["wasmer-compiler/wasmer-artifact-create"]
-static-artifact-load = ["wasmer-compiler/static-artifact-load"]
-static-artifact-create = ["wasmer-compiler/static-artifact-create"]
+wasmer-artifact-load = [
+  "wasmer-compiler/wasmer-artifact-load",
+  "wasmer-c-api-imports/wasmer-artifact-load",
+]
+wasmer-artifact-create = [
+  "wasmer-compiler/wasmer-artifact-create",
+  "wasmer-c-api-imports/wasmer-artifact-create",
+]
+static-artifact-load = [
+  "wasmer-compiler/static-artifact-load",
+  "wasmer-c-api-imports/static-artifact-load",
+]
+static-artifact-create = [
+  "wasmer-compiler/static-artifact-create",
+  "wasmer-c-api-imports/static-artifact-create",
+]
 webc_runner = ["virtual-fs", "webc"]
 # Deprecated features.
 jit = ["compiler"]
